+++ packages/sdks-tests/src/helpers/index.ts
@@ -37,6 +37,16 @@ async function screenshotOnFailure(
   }
 }
 
+const isIgnorableError = (error: Error) => {
+  return error.message.includes(
+    /**
+     * This error started appearing recently across all frameworks.
+     * It is most likely some playwright browser issue and not something we can fix in our code.
+     */
+    "Failed to execute 'observe' on 'PressureObserver': Access to the feature \"compute pressure\" is disallowed by permissions policy"
+  );
+};
+
 const test = base.extend<TestOptions>({
   // this is provided by `playwright.config.ts`
   packageName: ['DEFAULT' as any, { option: true }],
@@ -45,19 +55,23 @@ const test = base.extend<TestOptions>({
   ignoreHydrationErrors: [false, { option: true }],
   page: async ({ context, page, packageName, sdk, ignoreHydrationErrors }, use) => {
     if (packageName === ('DEFAULT' as any)) {
-      throw new Error('packageName is required');
+      throw new Error('`packageName` is required but was not provided.');
     }
     if (sdk === ('DEFAULT' as any)) {
-      throw new Error('sdk is required');
+      throw new Error('`sdk` is required but was not provided.');
     }
 
     context.on('weberror', err => {
+      if (isIgnorableError(err.error())) return;
+
       console.error(err.error());
-      throw new Error('Failing test due to error in browser: ' + err.error());
+      throw new Error('Test failed due to error thrown in browser: ' + err.error());
     });
     page.on('pageerror', err => {
+      if (isIgnorableError(err)) return;
+
       console.error(err);
-      throw new Error('Failing test due to error in browser: ' + err);
+      throw new Error('Test failed due to error thrown in browser: ' + err);
     });
 
     /**
@@ -80,7 +94,6 @@ const test = base.extend<TestOptions>({
         }
       });
     }
-
     if (sdk === 'angular') {
       page.on('console', msg => {
         const originalText = msg.text();
@@ -88,8 +101,20 @@ const test = base.extend<TestOptions>({
           throw new Error('Angular input not annotated error detected: ' + originalText);
         }
       });
+    } else if (sdk === 'vue') {
+      page.on('console', msg => {
+        const originalText = msg.text();
+        if (originalText.toLowerCase().includes('[vue warn]:')) {
+          throw new Error('Vue warning detected: ' + originalText);
+        }
+      });
+      context.on('console', msg => {
+        const originalText = msg.text();
+        if (originalText.toLowerCase().includes('[vue warn]:')) {
+          throw new Error('Vue warning detected: ' + originalText);
+        }
+      });
     }
-
     await use(page);
   },
 });
@@ -219,6 +244,7 @@ export const checkIfIsHydrationErrorMessage = (_text: string) => {
   const text = _text.toLowerCase();
   const isVueHydrationMismatch =
     text.includes('[vue warn]') && (text.includes('hydration') || text.includes('mismatch'));
+
   const isReactHydrationMismatch =
     text.includes('did not expect server') ||
     text.includes('content does not match') ||