+++ packages/sdks/output/nextjs/module-augmentations.d.ts
@@ -4,3 +4,7 @@
 declare module 'next/navigation' {
   export * from 'next/navigation.js';
 }
+
+declare module 'next/cache' {
+  export * from 'next/cache.js';
+}